fix VmDatabase tests
