+++ benches/Cargo.toml
@@ -14,8 +14,8 @@ ctrlc = "3.2.3"
 ed25519-dalek = { version = "2.0", features = ["rand_core"] }
 ethnum = "1.3"
 fuel-core = { path = "../crates/fuel-core", default-features = false, features = ["rocksdb-production"] }
-fuel-core-database = { path = "./../crates/database" }
 fuel-core-chain-config = { workspace = true }
+fuel-core-database = { path = "./../crates/database" }
 fuel-core-services = { path = "./../crates/services" }
 fuel-core-storage = { path = "./../crates/storage" }
 fuel-core-sync = { path = "./../crates/services/sync", features = ["benchmarking"] }
