+++ crates/fuel-core/Cargo.toml
@@ -41,6 +41,7 @@ hex = { version = "0.4", features = ["serde"] }
 hyper = { workspace = true }
 itertools = { workspace = true }
 postcard = { workspace = true, features = ["use-std"] }
+primitive-types = "0.12"
 rand = { workspace = true }
 rocksdb = { version = "0.21", default-features = false, features = [
     "lz4",
@@ -60,6 +61,7 @@ uuid = { version = "1.1", features = ["v4"] }
 
 [dev-dependencies]
 assert_matches = "1.5"
+fuel-core-executor = { workspace = true, features = ["test-helpers"] }
 fuel-core-services = { path = "./../services", features = ["test-helpers"] }
 fuel-core-storage = { path = "./../storage", features = ["test-helpers"] }
 fuel-core-trace = { path = "./../trace" }
@@ -68,7 +70,6 @@ mockall = { workspace = true }
 proptest = { workspace = true }
 test-case = { workspace = true }
 test-strategy = { workspace = true }
-fuel-core-executor = { workspace = true, features = ["test-helpers"] }
 
 [features]
 default = ["rocksdb"]
