+++ crates/fuel-core/src/lib.rs
@@ -36,6 +36,7 @@ pub mod state;
 
 // In the future this module will be a separate crate for `fuel-core-graphql-api`.
 mod graphql_api;
+pub mod vm_database;
 
 pub mod fuel_core_graphql_api {
     pub use crate::graphql_api::*;
